Update README.md
I created a tool named as Password strength checker in Python that is absolutely very helpful in checking the strength of passwords entered by users at various platforms. Password cracking is the most common way of hacking anyone's accounts which is why we need to be very careful while our setting passwords. Hence, in this project, we would see what type of passwords range from poor to strong.